// lib/data/seoul/types/culturalSpace.types.ts
export type SeoulSpaceRawRow = {
  FAC_NAME?: string; // 문화공간명
  SUBJCODE?: string; // 주제분류
  FAC_ADDR?: string; // 주소
  GUNAME?: string; // 자치구
  X_COORD?: string; // 경도
  Y_COORD?: string; // 위도
  PHNE?: string; // 전화번호
  FAX?: string; // 팩스
  HOMEPAGE?: string; // 홈페이지
  OPNHOURR?: string; // 관람시간
  ENTR_FEE?: string; // 입장료
  CLOSEDAY?: string; // 휴관일
  SEAT_CNT?: string; // 좌석수
  FAC_DESC?: string; // 시설설명
  MAIN_IMG?: string; // 대표이미지
};